+++ src/main/java/sugaEngine/graphics/ui/menu/Menu.java
@@ -1,5 +1,6 @@
-package sugaEngine.graphics;
+package sugaEngine.graphics.ui.menu;
 
+import sugaEngine.graphics.DrawListener;
 import sugaEngine.input.KeyValues;
 
 /**
